Move company cards related functions to CompanyCards file
